requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
rapidfuzz
markupsafe
sqlalchemy
asyncpg
cloudinary
aiohttp
passlib[bcrypt]
python-jose[cryptography]